requirements.txt
pandas
numpy
streamlit
gensim
st-star-rating
IMDbPY
